// dsl/packages/cli/test/test-specs/gen-out/client/typescript-client/src/model/EnumInlineRecord.ts
type ValueEnum = 'A' | 'B';
type Value_NullEnum = 'C' | 'D';
type Value_OptEnum = 'E' | 'F';
type Value_Opt_NullEnum = 'G' | 'H';
type ListEnum = 'A' | 'B';
type List_NullEnum = 'C' | 'D';
type List_Opt_NullEnum = 'G' | 'H';
// ...
export type EnumInlineRecord = {
	readonly value: ValueEnum;
	readonly value_Null: Value_NullEnum | null;
	readonly value_Opt?: Value_OptEnum;
	readonly value_Opt_Null?: Value_Opt_NullEnum | null;
	readonly list: ListEnum[];
	readonly list_Null: List_NullEnum[] | null;
	readonly list_Opt_Null?: List_Opt_NullEnum[] | null;
};